python/main.py
import sys
import os
import io
import json
from PIL import Image, ImageDraw

# Import local geometry logic
from geometry import calculate_grid_positions, calculate_spiral_positions

def mock_yolo_detect(image_path, target_class="book"):
    """
    Simulates YOLOv11 detection on the baseline image.
    Returns a list of bounding boxes and count.
    """
    print(f"🔍 Running YOLOv11 on {image_path}...")
    # In reality, load model(path)(image)
    # Here we mock the result of the 'failed' generation
    # Let's say it found 13 books instead of 16, messy positions
    detected_count = 13
    print(f"⚠️  Detection Result: Found {detected_count} '{target_class}' objects.")
    return detected_count

def correct_image_grid(image_path, output_path, rows=4, cols=4):
    """
    Pipeline to correct a Grid image
    """
    print("-" * 50)
    print("🛠️  STARTING CORRECTION PIPELINE: GRID 4x4")
    print("-" * 50)
    
    # 1. Detect
    count = mock_yolo_detect(image_path, "book")
    target_count = rows * cols
    
    if count != target_count:
        print(f"❌ FAIL: Expected {target_count}, found {count}. Initiating correction...")
        
    # 2. Calculate Truth (Symbolic Step)
    print("📐 Calculating Ground Truth positions with NumPy...")
    # Assume 1024x1024 image
    width, height = 1024, 1024
    cell_w = 200
    spacing = 20
    # Center grid
    grid_w = cols * (cell_w + spacing)
    grid_h = rows * (cell_w + spacing)
    top_left = ((width - grid_w)/2, (height - grid_h)/2)
    
    true_positions = calculate_grid_positions(top_left, rows, cols, cell_w, cell_w, spacing)
    print(f"✅ Calculated {len(true_positions)} centers.")
    
    # 3. Visualization (Debug Mask)
    print("🎨 Generating Correction Mask (ControlNet Input)...")
    mask = Image.new('RGB', (width, height), (0, 0, 0))
    draw = ImageDraw.Draw(mask)
    
    for i, (x,y) in enumerate(true_positions):
        # Draw expected boxes
        half = cell_w / 2
        box = [x-half, y-half, x+half, y+half]
        draw.rectangle(box, outline="white", width=4)
        # Draw number
        draw.text((x, y), str(i+1), fill="white", align="center")
        
    mask.save("debug_mask_grid.png")
    print("💾 Saved geometric truth to 'debug_mask_grid.png'")
    
    # 4. Inpainting / ControlNet Step (Simulation)
    print("🚀 Sending to Stable Diffusion [ControlNet: Canny/Depth]...")
    print("... (Processing on GPU) ...")
    print(f"✅ Image corrected and saved to {output_path}")

def main():
    # Example usage
    # Ensure you have the baseline images generated previously
    base_dir = "../examples"
    
    # 1. Correct Grid
    # We use the baseline we generated in the previous turn (conceptually)
    correct_image_grid("mock_input.png", "corrected_grid.png")

if __name__ == "__main__":
    main()
