python/requirements.txt
torch>=2.4.0
torchvision
diffusers>=0.30.0
transformers>=4.40.0
accelerate
opencv-python
numpy
pillow
scipy
gradio
ultralytics
controlnet_aux
